fix: Allow for .next to be a MouseInteraction and clean up its event handlers
We now remove the .drag event handlers, but, because multiple MouseInteractions
share the same DOM node and event name, this will remove all event handlers.

Therefore, if the 'parent' MouseInteraction reattaches its event handlers
when .next changes this is not an issue.

Also, if .next changes quickly, we might attach a stale interaction
because we awaited while in the meantime .next already changed.
This caused event handlers to be added, and never removed.